refactor(ucp-analytics): split smoke test into a sibling file
Moves all smoke-test code out of ucp_analytics.py into a new
ucp-analytics/smoke_test.py. The library file goes from ~1200 to
~800 lines and contains only the parser, BQWriter, UCPTracker,
SampleAgent, and the httpx event-hook factory — nothing that
exists solely to exercise the sample.

The smoke test (both stdout and --e2e BigQuery modes) keeps the
same surface and command shape; only the entrypoint script name
changes:

  python ucp_analytics.py            →  python smoke_test.py
  python ucp_analytics.py --e2e ...  →  python smoke_test.py --e2e ...

smoke_test.py imports the public surface (BQWriter, UCPTracker,
SampleAgent) from ucp_analytics and supplies its own
PrintWriter + CoverageTeeWriter + event generator + CLI. Behavior
is unchanged from PR #3 — 33 rows / 32 distinct event types in
stdout mode; e2e mode still refuses to report success when the
flush couldn't drain (verified against a bad project: still
exits 1 with the underlying BQ error).

Rationale: separating the smoke from the sample makes the "copy
this one file into your project" story literal again. The smoke
test is a regression check for this repo, not something a
consumer of the sample needs to carry. ucp_analytics.py is now
clean of CLI scaffolding, argparse, the synthetic event corpus,
the spec event-type manifest, and the verify-query polling
helper — none of which a downstream user needs.

Validated:
  python smoke_test.py                                  → 32/32
  python smoke_test.py --e2e --project-id REAL --verify → 32/32 + verify ok
  python smoke_test.py --e2e --project-id BAD           → exit 1, surfaces NotFound

README updated to point at smoke_test.py in both invocation
examples; root README's UCP Analytics bullet mentions the sibling
file. The on-disk line-count claim is updated (~800 lines).
